requirements.txt
pandas==0.24.2
matplotlib==3.0.3
seaborn==0.9.0
dfpvizpy==0.5.3
numpy==1.16.2
Pillow==6.0.0
elasticsearch==7.0.2
Flask==1.0.3
